Cube update
Refactored draw and update code into the Cube class so that cubes can be drawn independently. Also updated application so that a cube is drawn instead of a plane. Finally, updated Application to use the Cube class.

Instances of the Cube class are partially unique, but not entirely. So, this needs more to be fully refactored