Proper transparency added
Updated Material to determine if an object is transparent or not. The main draw loop has been updated to use this, as it draws all opaque objects first. Then it sorts transparent objects based on their distance from the camera and draws them. RasterState was also updated for this, as a third front face culling raster state was added as a member variable. Renderer was also updated, so if an object is transparent, it draws the backfaces and then the front faces of the object