Speed boost, debug camera and first person camera
Renamed the CameraController and FirstPersonCamera to DebugCameraController and DebugCamera. Added a FirstPersonController which moves the player in the same way as the debug camera, but the player y does not change. It also lets the player crouch by holding control. Lastly, added a speed boost to all cameras, so the player can hold shift to move faster